# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "领域知识个性化生成与多智能体协同决策系统",
    "description": "领域知识个性化生成与多智能体协同决策系统后端API\n\n## 认证方式\n- 除 `/auth/login`、`/auth/register`、健康检查外，所有接口需在请求头携带 `Authorization: Bearer <access_token>`\n- Token 通过 `POST /api/v1/auth/login` 获取，有效期由 `JWT_EXPIRE_MINUTES` 控制\n- Token 过期后通过 `POST /api/v1/auth/refresh` 刷新\n\n## 统一响应格式\n所有接口返回 `{code, message, data, timestamp}` 结构，`code=200` 表示成功",
    "contact": {
      "name": "知识系统团队",
      "url": "https://github.com/knowledge-system"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "当前环境（相对路径）"
    },
    {
      "url": "http://localhost:8000/api/v1",
      "description": "开发环境"
    },
    {
      "url": "https://staging.example.com/api/v1",
      "description": "预发布环境"
    },
    {
      "url": "https://api.example.com/api/v1",
      "description": "生产环境"
    }
  ],
  "paths": {
...
    "/api/v1/trainings": {
      "get": {
        "tags": [
          "企业培训"
        ],
        "summary": "获取培训任务列表",
        "description": "获取企业培训任务列表，支持分页、搜索、筛选",
        "operationId": "get_training_list_api_v1_trainings_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "training_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Training Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "企业培训"
        ],
        "summary": "创建培训任务",
        "description": "创建新的企业培训任务",
        "operationId": "create_training_api_v1_trainings_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trainings/stats/overview": {
      "get": {
        "tags": [
          "企业培训"
        ],
        "summary": "获取培训统计",
        "description": "获取培训统计概览（缓存 30 秒）",
        "operationId": "get_training_stats_api_v1_trainings_stats_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/trainings/transfers/list": {
      "get": {
        "tags": [
          "企业培训"
        ],
        "summary": "获取转岗培训列表",
        "description": "获取所有转岗培训记录",
        "operationId": "get_transfers_api_v1_trainings_transfers_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/trainings/skill-gaps/analysis": {
      "get": {
        "tags": [
          "企业培训"
        ],
        "summary": "获取技能差距分析",
        "description": "获取技能差距分析数据",
        "operationId": "get_skill_gaps_api_v1_trainings_skill_gaps_analysis_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "training_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Training Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trainings/batch-import": {
      "post": {
        "tags": [
          "企业培训"
        ],
        "summary": "批量导入培训任务",
        "description": "批量导入培训任务",
        "operationId": "batch_import_api_v1_trainings_batch_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingBatchImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/trainings/{training_id}": {
      "get": {
        "tags": [
          "企业培训"
        ],
        "summary": "获取培训任务详情",
        "description": "根据ID获取培训任务详情",
        "operationId": "get_training_detail_api_v1_trainings__training_id__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "training_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Training Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "企业培训"
        ],
        "summary": "更新培训任务",
        "description": "更新培训任务信息",
        "operationId": "update_training_api_v1_trainings__training_id__put",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "training_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Training Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "企业培训"
        ],
        "summary": "删除培训任务",
        "description": "删除培训任务",
        "operationId": "delete_training_api_v1_trainings__training_id__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "training_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Training Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TrainingBatchImportItem": {
        "properties": {
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "training_name": {
            "type": "string",
            "title": "Training Name"
          },
          "training_type": {
            "type": "string",
            "title": "Training Type",
            "default": "standard"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "participant_count": {
            "type": "integer",
            "title": "Participant Count",
            "default": 0
          },
          "responsible_person": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible Person"
          }
        },
        "type": "object",
        "required": [
          "company_name",
          "training_name"
        ],
        "title": "TrainingBatchImportItem",
        "description": "批量导入单条"
      },
      "TrainingBatchImportRequest": {
        "properties": {
          "trainings": {
            "items": {
              "$ref": "#/components/schemas/TrainingBatchImportItem"
            },
            "type": "array",
            "title": "Trainings"
          }
        },
        "type": "object",
        "required": [
          "trainings"
        ],
        "title": "TrainingBatchImportRequest",
        "description": "批量导入请求"
      },
      "TrainingCreate": {
        "properties": {
          "company_name": {
            "type": "string",
            "maxLength": 100,
            "title": "Company Name",
            "description": "企业名称"
          },
          "training_name": {
            "type": "string",
            "maxLength": 200,
            "title": "Training Name",
            "description": "培训名称"
          },
          "training_type": {
            "type": "string",
            "title": "Training Type",
            "description": "培训类型(standard/transfer)",
            "default": "standard"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "培训描述"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry",
            "description": "所属行业"
          },
          "modules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modules",
            "description": "培训模块列表"
          },
          "participant_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Participant Count",
            "description": "参与人数",
            "default": 0
          },
          "participants": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Participants",
            "description": "参与学员ID列表"
          },
          "responsible_person": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible Person",
            "description": "负责人"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date",
            "description": "开始日期"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date",
            "description": "结束日期"
          },
          "estimated_duration": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Estimated Duration",
            "description": "预计时长(天)",
            "default": 0
          },
          "is_transfer_training": {
            "type": "boolean",
            "title": "Is Transfer Training",
            "description": "是否转岗培训",
            "default": false
          },
          "transfer_from_position": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer From Position",
            "description": "原岗位"
          },
          "transfer_to_position": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer To Position",
            "description": "目标岗位"
          },
          "skill_gap_analysis": {
            "type": "object",
            "title": "Skill Gap Analysis",
            "description": "技能差距分析"
          }
        },
        "type": "object",
        "required": [
          "company_name",
          "training_name"
        ],
        "title": "TrainingCreate",
        "description": "创建培训任务"
      },
      "TrainingUpdate": {
        "properties": {
          "training_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "modules": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modules"
          },
          "participant_count": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Participant Count"
          },
          "participants": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Participants"
          },
          "responsible_person": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Responsible Person"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "estimated_duration": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Duration"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(planning|ongoing|completed|cancelled)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "progress_percentage": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Percentage"
          },
          "completed_modules": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Modules"
          },
          "is_transfer_training": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Transfer Training"
          },
          "transfer_from_position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer From Position"
          },
          "transfer_to_position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer To Position"
          },
          "skill_gap_analysis": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Gap Analysis"
          },
          "pass_rate": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pass Rate"
          },
          "average_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Score"
          },
          "satisfaction_rate": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Satisfaction Rate"
          }
        },
        "type": "object",
        "title": "TrainingUpdate",
        "description": "更新培训任务"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}